chore(@clayui/layout): update Layout component stories
